Feat(UI): migrate trips to query param routing and use trips GraphQL query
- Change /trips/[id] to /trips?id= (static-friendly routing)
- Replace transactionTags workaround in timeline with direct trips query
- Merge trip list and detail into single trips/+page.svelte
- Show 'View All →' button on timeline trips section instead of '+ New Trip'
- Add margin between trip cards
